requirements.txt
numpy==1.26.4
pandas==2.2.2
streamlit==1.33.0
google==3.0.0
google-api-core==2.18.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-cloud==0.34.0
google-cloud-aiplatform==1.48.0
google-cloud-bigquery==3.21.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
vertexai==1.48.0
protobuf>=3.20.3
openpyxl>=3.1.2
langchain>=0.0.347
langchain-google-vertexai>=1.0.3
faiss-cpu>=1.7.4
transformers>=4.40.1
pypdf>=4.2.0
langchain-community>=0.2.6
python-dotenv==1.0.1
